Create RawCheckBox and change CheckBox
